api/requirements.txt
pydantic
fastapi
pandas
uvicorn
mlflow
boto3
python-dotenv
xgboost==1.7.6